+++ src/main/java/com/project/moflis/location/repository/LocationRepository.java
@@ -3,9 +3,9 @@ package com.project.moflis.location.repository;
 import com.project.moflis.location.entity.Location;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocationRepository extends JpaRepository<Location, Integer> {
+public interface LocationRepository extends JpaRepository<Location, Long> {
 
-    Location findByUserId(Integer userId);
+    Location findByUserId(Long userId);
 
-    boolean existsByUserId(Integer userId);
+    boolean existsByUserId(Long userId);
 }
