+++ src/main/java/com/project/moflis/profile/command/UpdateProfileCommand.java
@@ -4,10 +4,10 @@ import lombok.Getter;
 
 @Getter
 public class UpdateProfileCommand {
-    private final Integer userId;
+    private final Long userId;
     private final String intro;
 
-    public UpdateProfileCommand(Integer userId, String intro) {
+    public UpdateProfileCommand(Long userId, String intro) {
         this.userId = userId;
         this.intro = intro;
     }
