+++ src/main/java/com/project/moflis/profile/repository/ProfileRepository.java
@@ -3,8 +3,8 @@ package com.project.moflis.profile.repository;
 import com.project.moflis.profile.entity.Profile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProfileRepository extends JpaRepository<Profile, Integer> {
+public interface ProfileRepository extends JpaRepository<Profile, Long> {
 
-    Profile findByUserId(Integer userId);
+    Profile findByUserId(Long userId);
 
 }
