+++ src/main/java/com/project/moflis/recurringschedules/dto/RecurringScheduleResponse.java
@@ -12,8 +12,8 @@ import java.time.LocalTime;
 @Setter
 public class RecurringScheduleResponse {
 
-    private Integer id;
-    private Integer userId;
+    private Long id;
+    private Long userId;
     private RepeatType repeatType;
     private LocalDate startDate;
     private LocalDate endDate;
