+++ src/test/java/com/project/moflis/command/ProfileCommandTest.java
@@ -9,7 +9,7 @@ class ProfileCommandTest {
 
     @Test
     void ProfileCommand() {
-        Integer userId = 1;
+        Long userId = 1L;
         String intro = "이것은 테스트";
 
         AddProfileCommand command = new AddProfileCommand(userId, intro);
